+++ projects/core/src/main/kotlin/site/siredvin/peripheralworks/xplat/ModBlocksReference.kt
@@ -4,17 +4,17 @@ import net.minecraft.world.level.block.Block
 
 interface ModBlocksReference {
     companion object {
-        private var _IMPL: ModBlocksReference? = null
+        private var impl: ModBlocksReference? = null
 
         fun configure(impl: ModBlocksReference) {
-            _IMPL = impl
+            this.impl = impl
         }
 
         fun get(): ModBlocksReference {
-            if (_IMPL == null) {
+            if (impl == null) {
                 throw IllegalStateException("You should init PeripheralWorks Platform first")
             }
-            return _IMPL!!
+            return impl!!
         }
     }