+++ projects/fabric/src/main/kotlin/site/siredvin/peripheralworks/integrations/team_reborn_energy/Configuration.kt
@@ -5,27 +5,27 @@ import site.siredvin.peripheralium.api.config.IConfigHandler
 
 object Configuration : IConfigHandler {
 
-    private const val defaultEnergyToFuelRate = 50 // really calculated :) Trust me
+    private const val DEFAULT_ENERGY_TO_FUEL_RATE = 50 // really calculated :) Trust me
 
-    private var ENABLE_ENERGY_STORAGE: ForgeConfigSpec.BooleanValue? = null
-    private var ENABLE_TURTLE_REFUEL_WITH_ENERGY: ForgeConfigSpec.BooleanValue? = null
-    private var ENERGY_TO_FUEL_RATE: ForgeConfigSpec.IntValue? = null
+    private var enableEnergyStorageConfig: ForgeConfigSpec.BooleanValue? = null
+    private var enableTurtleRefuelWithEnergyConfig: ForgeConfigSpec.BooleanValue? = null
+    private var energyToFuelRateConfig: ForgeConfigSpec.IntValue? = null
 
     val enableEnergyStorage: Boolean
-        get() = ENABLE_ENERGY_STORAGE?.get() ?: true
-    val enableTurtleRefulWithEnergy: Boolean
-        get() = ENABLE_TURTLE_REFUEL_WITH_ENERGY?.get() ?: true
+        get() = enableEnergyStorageConfig?.get() ?: true
+    val enableTurtleRefuelWithEnergy: Boolean
+        get() = enableTurtleRefuelWithEnergyConfig?.get() ?: true
     val energyToFuelRate: Int
-        get() = ENERGY_TO_FUEL_RATE?.get() ?: defaultEnergyToFuelRate
+        get() = energyToFuelRateConfig?.get() ?: DEFAULT_ENERGY_TO_FUEL_RATE
 
     override val name: String
         get() = "team_reborn_energy"
 
     override fun addToConfig(builder: ForgeConfigSpec.Builder) {
-        ENABLE_ENERGY_STORAGE = builder.comment("Enables energy storage integration").define("enableEnergyStorage", true)
-        ENABLE_TURTLE_REFUEL_WITH_ENERGY = builder.comment("Enables turtle refueling with items with energy")
+        enableEnergyStorageConfig = builder.comment("Enables energy storage integration").define("enableEnergyStorage", true)
+        enableTurtleRefuelWithEnergyConfig = builder.comment("Enables turtle refueling with items with energy")
             .define("enableTurtleRefuelWithEnergy", true)
-        ENERGY_TO_FUEL_RATE = builder.comment("Controls how many energy required for one fuel point")
-            .defineInRange("energyToFuelRate", defaultEnergyToFuelRate, 1, Int.MAX_VALUE)
+        energyToFuelRateConfig = builder.comment("Controls how many energy required for one fuel point")
+            .defineInRange("energyToFuelRate", DEFAULT_ENERGY_TO_FUEL_RATE, 1, Int.MAX_VALUE)
     }
 }