+++ projects/fabric/src/main/kotlin/site/siredvin/peripheralworks/integrations/universal_shops/Configuration.kt
@@ -5,16 +5,16 @@ import site.siredvin.peripheralium.api.config.IConfigHandler
 
 object Configuration : IConfigHandler {
 
-    private var ENABLE_SHOPS: ForgeConfigSpec.BooleanValue? = null
+    private var enableShopsConfig: ForgeConfigSpec.BooleanValue? = null
 
     val enableShops: Boolean
-        get() = ENABLE_SHOPS?.get() ?: true
+        get() = enableShopsConfig?.get() ?: true
 
     override val name: String
         get() = "universal_shops"
 
     override fun addToConfig(builder: ForgeConfigSpec.Builder) {
-        ENABLE_SHOPS = builder.comment("Enables shops integration")
+        enableShopsConfig = builder.comment("Enables shops integration")
             .define("enableShops", true)
     }
 }