+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/easy_villagers/Configuration.kt
@@ -5,15 +5,15 @@ import site.siredvin.peripheralium.api.config.IConfigHandler
 
 object Configuration : IConfigHandler {
 
-    private var ENABLE_AUTO_TRADER: ForgeConfigSpec.BooleanValue? = null
+    private var enableAutoTradeConfig: ForgeConfigSpec.BooleanValue? = null
 
     val enableAutoTrader: Boolean
-        get() = ENABLE_AUTO_TRADER?.get() ?: true
+        get() = enableAutoTradeConfig?.get() ?: true
 
     override val name: String
         get() = "easy_villagers"
 
     override fun addToConfig(builder: ForgeConfigSpec.Builder) {
-        ENABLE_AUTO_TRADER = builder.comment("Enables auto trader integration").define("enableAutoTrader", true)
+        enableAutoTradeConfig = builder.comment("Enables auto trader integration").define("enableAutoTrader", true)
     }
 }