+++ projects/forge/src/main/kotlin/site/siredvin/peripheralworks/integrations/integrateddynamics/Configuration.kt
@@ -5,20 +5,20 @@ import site.siredvin.peripheralium.api.config.IConfigHandler
 
 object Configuration : IConfigHandler {
 
-    private var ENABLE_VARIABLE_STORE: ForgeConfigSpec.BooleanValue? = null
-    private var ENABLE_COMPUTER_ASPECT: ForgeConfigSpec.BooleanValue? = null
+    private var enableVariableStoreConfig: ForgeConfigSpec.BooleanValue? = null
+    private var enableComputerAspectConfig: ForgeConfigSpec.BooleanValue? = null
 
     val enableVariableStore: Boolean
-        get() = ENABLE_VARIABLE_STORE?.get() ?: true
+        get() = enableVariableStoreConfig?.get() ?: true
 
     val enableComputerAspect: Boolean
-        get() = ENABLE_COMPUTER_ASPECT?.get() ?: true
+        get() = enableComputerAspectConfig?.get() ?: true
 
     override val name: String
         get() = "integrateddynamics"
 
     override fun addToConfig(builder: ForgeConfigSpec.Builder) {
-        ENABLE_VARIABLE_STORE = builder.comment("Enables variable store integration").define("enableVariableStore", true)
-        ENABLE_COMPUTER_ASPECT = builder.comment("Enables computer aspect for block reader").define("enableComputerAspect", true)
+        enableVariableStoreConfig = builder.comment("Enables variable store integration").define("enableVariableStore", true)
+        enableComputerAspectConfig = builder.comment("Enables computer aspect for block reader").define("enableComputerAspect", true)
     }
 }