+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `registryID` added to `describe` function of informative registry
+
+### Fixed
+
+- Recipe registry is usable again (by [Janrupf](https://github.com/Janrupf))
+- Cooldowns can be set to 0 now
+- `getPatternsFor` Double/Int handling
+
 ## [1.4.3] - 2024-02-01
 
 ### Changed